renamed the card to card pic for consistency. added a HEAD tag to the main page to read va soft wash
